+++ seantis/agencies/upgrades.py
@@ -5,6 +5,7 @@ from zope.component import getUtility
 from plone import api
 from plone.dexterity.interfaces import IDexterityFTI
 
+from seantis.people import catalog_id
 from seantis.people.supermodel.indexing import update_related_indexes
 
 
@@ -41,3 +42,20 @@ def upgrade_1001_to_1002(context):
     upgrade_portal_type(
         'seantis.agencies.membership', 'seantis.agencies', 'default'
     )
+
+
+def upgrade_1002_to_1003(context):
+    upgrade_portal_type(
+        'seantis.agencies.organization', 'seantis.agencies', 'default'
+    )
+
+
+def upgrade_1003_to_1004(context):
+    upgrade_portal_type(
+        'seantis.agencies.organization', 'seantis.agencies', 'default'
+    )
+
+
+def upgrade_1004_to_1005(context):
+    catalog = api.portal.get_tool(catalog_id)
+    catalog.refreshCatalog(clear=1)